report : use lint
 npm run lint

+++ contracts/connection/TokamakStaker.sol
@@ -9,19 +9,34 @@ import "@openzeppelin/contracts/access/AccessControl.sol";
 import "../stake/StakeTONStorage.sol";
 
 interface IERC20BASE {
-    function totalSupply() external view returns (uint);
-    function balanceOf(address owner) external view returns (uint);
+    function totalSupply() external view returns (uint256);
+
+    function balanceOf(address owner) external view returns (uint256);
+
     //function allowance(address owner, address spender) external view returns (uint);
-    function approve(address spender, uint value) external returns (bool);
-    function transfer(address to, uint value) external returns (bool);
-    function transferFrom(address from, address to, uint value) external returns (bool);
+    function approve(address spender, uint256 value) external returns (bool);
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function transferFrom(
+        address from,
+        address to,
+        uint256 value
+    ) external returns (bool);
 }
+
 interface ITokamakRegistry {
     function getTokamak()
         external
         view
-        returns (address,address,address,address);
+        returns (
+            address,
+            address,
+            address,
+            address
+        );
 }
+
 /// @title The connector that integrates zkopru and tokamak
 contract TokamakStaker is StakeTONStorage, AccessControl {
     using SafeMath for uint256;
@@ -61,12 +76,15 @@ contract TokamakStaker is StakeTONStorage, AccessControl {
 
     event tokamakStaked(address layer2, uint256 amount);
     event tokamakRequestedUnStaking(address layer2, uint256 amount);
-    event tokamakProcessedUnStaking(address layer2, uint256 rn, bool receiveTON);
+    event tokamakProcessedUnStaking(
+        address layer2,
+        uint256 rn,
+        bool receiveTON
+    );
 
-    function setRegistryNDefi(
-        address _registry,
-        address _defiAddr
-    ) external onlyOwner
+    function setRegistryNDefi(address _registry, address _defiAddr)
+        external
+        onlyOwner
         nonZero(_registry)
     {
         stakeRegistry = _registry;
@@ -117,35 +135,54 @@ contract TokamakStaker is StakeTONStorage, AccessControl {
     }
 
     function tokamakStaking(address _layer2)
-        external lock nonZero(stakeRegistry) nonZero(_layer2)
+        external
+        lock
+        nonZero(stakeRegistry)
+        nonZero(_layer2)
     {
         require(block.number <= endBlock, "period end");
-        require(
-            IIStake1Vault(vault).saleClosed() == true,
-            "not closed"
-        );
-        defiStatus = uint(LibTokenStake1.DefiStatus.DEPOSITED);
+        require(IIStake1Vault(vault).saleClosed() == true, "not closed");
+        defiStatus = uint256(LibTokenStake1.DefiStatus.DEPOSITED);
 
-        (address ton, address wton, address depositManager, address seigManager) = ITokamakRegistry(stakeRegistry).getTokamak();
-        require(ton != address(0) && wton != address(0) && depositManager != address(0) && seigManager != address(0),
+        (
+            address ton,
+            address wton,
+            address depositManager,
+            address seigManager
+        ) = ITokamakRegistry(stakeRegistry).getTokamak();
+        require(
+            ton != address(0) &&
+                wton != address(0) &&
+                depositManager != address(0) &&
+                seigManager != address(0),
             "ITokamakRegistry zero"
         );
 
-        uint256 globalWithdrawalDelay = IIDepositManager(depositManager).globalWithdrawalDelay();
-        require(block.number < endBlock - globalWithdrawalDelay, "period(withdrawalDelay) end");
+        uint256 globalWithdrawalDelay =
+            IIDepositManager(depositManager).globalWithdrawalDelay();
+        require(
+            block.number < endBlock - globalWithdrawalDelay,
+            "period(withdrawalDelay) end"
+        );
 
         uint256 _amount = IERC20BASE(ton).balanceOf(address(this));
-        require(
-                _amount > 0,
-            "zero"
-        );
+        require(_amount > 0, "zero");
 
         if (tokamakLayer2 == address(0)) tokamakLayer2 = _layer2;
         else {
-
-            if (IISeigManager(seigManager).stakeOf(tokamakLayer2, address(this)) > 0 ||
-                IIDepositManager(depositManager).pendingUnstaked( tokamakLayer2, address(this)) > 0) {
-                require( tokamakLayer2 == _layer2, "different layer" );
+            if (
+                IISeigManager(seigManager).stakeOf(
+                    tokamakLayer2,
+                    address(this)
+                ) >
+                0 ||
+                IIDepositManager(depositManager).pendingUnstaked(
+                    tokamakLayer2,
+                    address(this)
+                ) >
+                0
+            ) {
+                require(tokamakLayer2 == _layer2, "different layer");
             } else {
                 if (tokamakLayer2 != _layer2) tokamakLayer2 == _layer2;
             }
@@ -161,27 +198,33 @@ contract TokamakStaker is StakeTONStorage, AccessControl {
     }
 
     function tokamakRequestUnStaking(address _layer2, uint256 wtonAmount)
-        public lock nonZero(stakeRegistry) sameTokamakLayer(_layer2)
+        public
+        lock
+        nonZero(stakeRegistry)
+        sameTokamakLayer(_layer2)
     {
-        require(
-            IIStake1Vault(vault).saleClosed() == true,
-            "not closed"
-        );
-        defiStatus = uint(LibTokenStake1.DefiStatus.REQUESTWITHDRAW);
+        require(IIStake1Vault(vault).saleClosed() == true, "not closed");
+        defiStatus = uint256(LibTokenStake1.DefiStatus.REQUESTWITHDRAW);
         requestNum = requestNum.add(1);
 
-        (address ton, address wton, address depositManager, address seigManager) = ITokamakRegistry(stakeRegistry).getTokamak();
-        require(ton != address(0) && wton != address(0) && depositManager != address(0) && seigManager != address(0),
+        (
+            address ton,
+            address wton,
+            address depositManager,
+            address seigManager
+        ) = ITokamakRegistry(stakeRegistry).getTokamak();
+        require(
+            ton != address(0) &&
+                wton != address(0) &&
+                depositManager != address(0) &&
+                seigManager != address(0),
             "ITokamakRegistry zero"
         );
 
         uint256 stakeOf =
             IISeigManager(seigManager).stakeOf(_layer2, address(this));
 
-        require(
-            stakeOf >= wtonAmount,
-            "lack"
-        );
+        require(stakeOf >= wtonAmount, "lack");
 
         IIDepositManager(depositManager).requestWithdrawal(_layer2, wtonAmount);
 
@@ -189,34 +232,49 @@ contract TokamakStaker is StakeTONStorage, AccessControl {
     }
 
     function tokamakProcessUnStaking(address _layer2, bool receiveTON)
-        public lock nonZero(stakeRegistry) sameTokamakLayer(_layer2)
+        public
+        lock
+        nonZero(stakeRegistry)
+        sameTokamakLayer(_layer2)
     {
         require(
-            defiStatus != uint(LibTokenStake1.DefiStatus.WITHDRAW),
+            defiStatus != uint256(LibTokenStake1.DefiStatus.WITHDRAW),
             "Already ProcessUnStaking"
         );
-        require(
-            IIStake1Vault(vault).saleClosed() == true,
-            "not closed"
-        );
-        defiStatus = uint(LibTokenStake1.DefiStatus.WITHDRAW);
+        require(IIStake1Vault(vault).saleClosed() == true, "not closed");
+        defiStatus = uint256(LibTokenStake1.DefiStatus.WITHDRAW);
         uint256 rn = requestNum;
         requestNum = 0;
 
-        (address ton, address wton, address depositManager, address seigManager) = ITokamakRegistry(stakeRegistry).getTokamak();
-        require(ton != address(0) && wton != address(0) && depositManager != address(0) && seigManager != address(0),
+        (
+            address ton,
+            address wton,
+            address depositManager,
+            address seigManager
+        ) = ITokamakRegistry(stakeRegistry).getTokamak();
+        require(
+            ton != address(0) &&
+                wton != address(0) &&
+                depositManager != address(0) &&
+                seigManager != address(0),
             "ITokamakRegistry zero"
         );
 
-        if (IISeigManager(seigManager).stakeOf(tokamakLayer2, address(this)) == 0 ) tokamakLayer2 = address(0);
+        if (
+            IISeigManager(seigManager).stakeOf(tokamakLayer2, address(this)) ==
+            0
+        ) tokamakLayer2 = address(0);
 
         fromTokamak += IIDepositManager(depositManager).pendingUnstaked(
-                _layer2,
-                address(this)
-            );
-        IIDepositManager(depositManager).processRequests(_layer2, rn, receiveTON);
+            _layer2,
+            address(this)
+        );
+        IIDepositManager(depositManager).processRequests(
+            _layer2,
+            rn,
+            receiveTON
+        );
 
         emit tokamakProcessedUnStaking(_layer2, rn, receiveTON);
     }
-
 }
