+++ contracts/stake/StakeTONStorage.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.7.6;
 import "./Stake1Storage.sol";
 
 contract StakeTONStorage is Stake1Storage {
-
     address public stakeRegistry;
     address public tokamakLayer2;
     address internal _uniswapRouter;
@@ -14,7 +13,7 @@ contract StakeTONStorage is Stake1Storage {
     uint256 public fromTokamak;
     uint256 public toUniswapTON;
     uint256 public swappedAmountFLD;
-    uint public defiStatus;
+    uint256 public defiStatus;
 
     uint256 public requestNum;
 }
